#86976a177[DONE] Prices conversion implementation

+++ mm-expenses-manager-finance/app/src/main/java/mm/expenses/manager/finance/exchangerate/provider/nbp/NbpClient.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collection;
 import java.util.stream.Collectors;
@@ -38,7 +39,7 @@ interface NbpClient {
     @JsonIgnoreProperties(ignoreUnknown = true)
     class TableRateDto {
         private String code;
-        private Double mid;
+        private BigDecimal mid;
     }
 
     @Data
